LayoutTests/media/unsupported-rtsp-expected.txt

Test that QuickTime file with RTSP URL generates a load error.

EVENT(error)
EXPECTED (video.error != 'null') OK
EXPECTED (video.error.code == '2') OK
END OF TEST
